python_backend_agent_scaffold/requirements.txt
livekit-agents>=1.0.0
livekit-plugins-openai>=1.0.0
livekit-plugins-azure>=1.0.0
python-dotenv>=0.20.0
aiohttp>=3.8.0
